+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.8.0 (in progress)
 -------------------
 * **(breaking change)** `find()` now distinguishes between files and directories and by default searches only for files (previously searched for both).
+* **DEPRECATION:** Don't do `jetpack.read('sth', 'buf')`, do `jetpack.read('sth', 'buffer')` instead.
 
 0.7.3 (2016-03-21)
 -------------------